♻️ Virtual NodeをSVGに変換する処理を分離
Virtual NodeをSVGに変換する関数 vnode2str を、src/compiler/vnode-to-str.mjs ファイルに移動し、外部ファイル化した。
